Return title instead of label in Goobi XML
Fixes sul-dlss/goobi-robot#30

When the Goobi XML is created, use the text of the first MODS title for the document title if present. If not present, use the object label. Prior to this change, the object label was always used in this context, but the label is not always present, requiring downstream failure and manual remediation.