Remove unnecessary tzinfo-data dependency
This is only included in the bundle on environments we do not deploy in (Windows and JRuby), and in the meantime, its presence in the bundle causes every bundle command to include a warning message. This message can trip up shell aliases and functions that use bundler, such as grepping the bundle to sleuth out behavior within dependencies.

Refs sul-dlss/argo#1277