Add ActiveRecord and Postgres as dependencies
Connects to #336

This makes it possible to reliably stash information about background jobs in a persistent store, which will be used to run virtual object creation in the background.

Includes `factory_bot_rails` as a dependency which will make it easier to spin
up fixture-like factories when testing with database entities later.