Disable writing to solr
This mimics what Argo is doing and allows dor_indexing_app to handle all indexing. This is an important step as the indexing code is no longer shared in dor-services gem, so they could potentially be indexing different things.  This also helps reduce load on solr, as every write to fedora already triggers an indexing action via a message from Fedora 3 to activemq, which is read by dor_indexing_app. Furthermore, this increases the reliability of dor-services-app as we no longer have a dependency on writing to solr regularly.

Fixes #30