Store administrative tags in the database
Connects to #679

This change allows reading administrative tags from the database OR Fedora, and only writes them TO the database. When tags are written to the database for a druid that currently has no rows present, tags will be read from Fedora and written to the database. This will give us some time to complete the migration of tags from Fedora to the database, making a gradual transition

Also includes:
* Return HTTP 409 Conflict if client asks to duplicate tag/druid rows
* Add ability to update, destroy, and replace administrative tags
* Add content_type sniffing (used only internally in dor-services-app, so not exposed via API)
* Use AdministrativeTags service instead of dor-services #tags method (dog-fooding)
* Prefer running tests within a transaction to using database_cleaner