Differentiate between env-specific database names
This prevents Rails from warning the developer about using the same database for both dev and test, which most of us developers are straddling frequently. This does not affect deployed environments because the checked-in config/database.yml is overwritten with a file that is managed via Puppet. This is intended to remove drag on development.