Run the docker image in RAILS_ENV=production
Avoids this concurrency error https://github.com/rails/rails/issues/32451
This happens when Argo registers an object and dor-services-app calls dor-indexing-app
which calls back to dor-services-app for the list of AdministrativeTags.
dor-services-app cannot respond to this second request, so the indexing call times out.
This probably wouldn't be a problem in Rails 6, but ActiveFedora is preventing that upgrade.