Bootstrap the Ur-AdminPolicy
There is no way to create new objects in the system until the Ur-AdminPolicy exists.  Presently Argo has a factory that writes directly into Fedora to accomplish this, but we want to decouple Argo from Fedora.  Now, we can turn on a feature flag when this is run in a test container which will automatically bootstrap the Ur-AdminPolicy.