Ensure use and license fields are initialized before setting them
I honestly am not sure if we would ever see this in production but I have data that looks like this in Argo locally and I figured it was worth writing a new test and patching it in. Not sure what the harm is, despite the small cost of added code.

Connects to sul-dlss/argo#2404