Fix sidekiq web UI
Fixes #2638

This PR follows the instructions from the Sidekiq docs: https://github.com/mperham/sidekiq/wiki/Monitoring#rails-api-application-session-configuration

It also adds a SECRET_KEY_BASE to the web docker-compose service, which is required now that we switched to RAILS_ENV=production. This value is not used anywhere else, so it seems safe to have here. No .env hijinks needed, I do not think.