Match Cocina structures as JSON but give hashes to SuperDiff
The `equal_cocina_model` matcher compares a JSON string as actual value against a Cocina model coerced to JSON as expected value. We want to compare these as JSON rather than as hashes, else we'll have to start deep-converting some values within the hashes, thinking of date/time values in particular. This matching behavior continues what dor-services-app was already doing before this custom matcher was written.

Note, though, that when actual and expected values do *not* match, we coerce both values to hashes to allow the `super_diff` gem to highlight the areas that differ. This is easier to scan than two giant JSON strings.