Fix bug with rights determinations when metadata has multiple access nodes
Connects to sul-dlss/argo#2382

Rights metadata containing e.g. file-specific nodes reveal a bug in the prior `FromFedora/Access` implementation, namely the assumption that the presence of a single access=read node is meaningful. Instead of hand-crafted, buggy XPaths, re-use the logic that is already in place in the dor-rights-auth gem.

This PR adds full object-level rights mappings on both the to- and from-fedora sides (32 cases a piece).