Prefer passing nokogiri docs (vs. strings) to equivalent XML
Why? Because calling ds.ng_xml.to_xml produces a full document string including the XML declaration... and when you pass that string to EquivalentXml, it compares it as an XML fragment which when passed an XML string with a declaration, it proceeds to lose XML entities like &amp; which is why the location:m&m specs were failing and exhibiting weird prior behavior. I cannot explain that Nokogiri behavior but in the meantime we can use EquivalentXml better.