Switch roundtrip validation failure HTTP status to unprocessable_entity
This allows SDR-API to not have to parse the conflict error differently for different types of errors. See https://github.com/sul-dlss/sdr-api/issues/415.

Furthermore 412 is a better fit, as there is no value the user could change to bring it out of this conflict state.  Finally, UnexpectedBuildError already uses 412, and this makes the RoundtripValidationError consistent with that.