Republish collection members when collection is (re-)published
Fixes #4195

This commit changes the `Publish::MetadataTransferService` such that when a collection is (re-)published, its members are (re-)published. Not all members, though: only those members that are 1) already published, and 2) not in the `Opened` state. It does this work using the existing `MemberService` which is extended here to add a param excluding items in the `Opened` state.