Mark DOI as registered when object is in graveyard APO
Fixes sul-dlss/happy-heron#2750

This commit adds a new check to the class that maps Cocina objects to Datacite attributes: if the object is in the SDR Graveyard APO, use the `register` event instead of the `publish` event. This has the effect of making sure DOIs of decommissioned (soft-deleted) objects are no longer findable.