Fix bug with generating thumbnail links for virtual objects
Fixes #4456

This bug, as far as I can tell, has been around since we cocinized the thumbnail service which prior to this commit returned the druid of the first member of a virtual object as its thumbnail link. Instead, pass this druid back to the thumbnail service to get the expected thumbnail link (the bare druid followed by the filename of the first jp2 of the first member).