Retry Folio edit operations a few times
Fixes #4454

If multiple Folio edits are requested serially, at least two different exceptions may be raised:

* a Faraday timeout
* a StandardError emitted from FolioClient

This is documented in the FolioClient gem: https://github.com/sul-dlss/folio_client/blob/181ee68c8879030734f5ad6a639904145c4ffc06/lib/folio_client/records_editor.rb#L19-L24

Both of these may be safely retried a small number of times, and both conditions go away within seconds. I can be convinced that we ought to transparently handle this in the gem, but I'm not sure we'll be encountered this outside the DSA FolioWriter so maybe this is a fine location for now.