Add catalog limits
This commit includes options for generating stats for Cocina objects
that have links to the catalog, or don't have links to the catalog. This
is to try to help see how the metadata changes based on whether it was
derived from MARC or not.

This was accidentally not committed before because it was hacked on live
over on the sdr-infra server (sorry!)