Normalize administrative tag labels to remove leading/trailing whitespace
Fixes sul-dlss/argo#4327

The regular expression used in the DSA codebase and at the API layer (as expressed in `openapi.yml`) allows the user to enter both leading and trailing whitespace for administrative tags, and we don't want this whitespace persisted. The solution in this commit adds normalization at the DB model layer, removing the bad whitespace.

I opted for this solution over a validation-based solution (tightening the various regular expressions) because that may make remediation of currently persisted bad data harder.

We could consider a follow-up step that tweaks the regexes, but also, that's arguably of limited value once this commit is merged and bad data is remediated. Why? More work than it's worth, because we'd likely want to couple that change with user messaging to give them more guidance about how to enter a valid admin tag.