Add a migration service creating repository objects & versions
Fixes #4817

This commit adds a new `RepositoryObjectMigrator` service that takes a single druid/external ID---for a DRO or collection or admin policy---as input and creates a repository object instance and one or more repository object version instances. It does all of its work in a transaction and migrates all relevant information from the `ObjectVersion` table and the `Dro|Collection|AdminPolicy` table.